Changed startup file
