+++ .vscode/launch.json
@@ -8,7 +8,7 @@
             "name": "Python: Current File",
             "type": "python",
             "request": "launch",
-            "program": "src/index.py",
+            "program": "src/app.py",
             "console": "integratedTerminal",
         }
     ]
