+++ src/view/home.py
@@ -1,16 +1,15 @@
 import dash_bootstrap_components as dbc
 import plotly.express as px
-from dash import Input, Output, dcc, html
+from dash import Input, Output, callback, dcc, html
 from dash.exceptions import PreventUpdate
 
-from app import app
 from service.account_transactions import get_report
 from utils.functions import formatter_currency_with_cents
 
 layout = html.Div(id="home_content")
 
 
-@app.callback(
+@callback(
     Output("home_content", "children"),
     Input("url", "href"),
 )
