+++ src/view/income_finder.py
@@ -1,10 +1,9 @@
 import dash_bootstrap_components as dbc
 import dash_tabulator
-from dash import dcc, html
+from dash import callback, dcc, html
 from dash.dependencies import Input, Output, State
 from dash.exceptions import PreventUpdate
 
-from app import app
 from service.chart_helper import update_graph
 from service.search_income import watchlist_income
 from utils.constants import screener_list
@@ -187,7 +186,7 @@ layout = dbc.Form(
 )
 
 
-@app.callback(
+@callback(
     [
         Output("income-output", "children"),
         Output("income-message", "is_open"),
@@ -292,7 +291,7 @@ def on_button_click(
 
 # dash_tabulator can register a callback on rowClicked
 # to receive a dict of the row values
-@app.callback(
+@callback(
     [
         Output("chart-output", "children"),
         Output("modal-chart", "is_open"),
